fix: Fixed the improper display of elements after a new gif is added
When a new gif was added sometime the multiple layout was happening multiple times and overwriting the gif that was being displayed with nothing.
Now when I add a gif to the collection it is inserted at the end. When the CollectionView needs to get that data it queries the model thru it's own internal functions.
This allows us to take out all the reloadData commands too.